+++ packages/@livestore/common/src/schema/state/sqlite/mod.ts
@@ -3,20 +3,11 @@ import { shouldNeverHappen } from '@livestore/utils'
 import type { MigrationOptions } from '../../../adapter-types.ts'
 import type { Materializer } from '../../EventDef/mod.ts'
 import type { InternalState } from '../../schema.ts'
-import { ClientDocumentTableDefSymbol, tableIsClientDocumentTable } from './client-document-def.ts'
 import { SqliteAst } from './db-schema/mod.ts'
 import { stateSystemTables } from './system-tables/state-tables.ts'
 import type { TableDef, TableDefBase } from './table-def.ts'
 
 export * from '../../EventDef/mod.ts'
-export {
-  type ClientDocumentTableDef,
-  ClientDocumentTableDefSymbol,
-  type ClientDocumentTableOptions,
-  clientDocument,
-  createOptimisticEventSchema,
-  tableIsClientDocumentTable,
-} from './client-document-def.ts'
 export * from './column-annotations.ts'
 export * from './column-spec.ts'
 export * from './table-def.ts'
@@ -44,15 +35,6 @@ export const makeState = <TStateInput extends InputState>(inputSchema: TStateInp
 
   for (const [name, materializer] of Object.entries(inputSchema.materializers)) {
     materializers.set(name, materializer)
-  }
-
-  for (const tableDef of inputTables) {
-    if (tableIsClientDocumentTable(tableDef) === true) {
-      materializers.set(
-        tableDef[ClientDocumentTableDefSymbol].derived.setEventDef.name,
-        tableDef[ClientDocumentTableDefSymbol].derived.setMaterializer,
-      )
-    }
   }
 
   const hash = SqliteAst.hash({